backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
asyncpg==0.30.0
greenlet==3.3.2
redis==5.2.1
qdrant-client==1.12.1
sentence-transformers==3.3.1
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.13.1
httpx==0.28.1
groq==1.0.0
numpy==2.1.3
langgraph==1.0.10
PyJWT==2.9.0
scikit-learn==1.5.2
prometheus-client==0.21.0
authlib==1.3.2
websockets==14.1
python-multipart==0.0.19
passlib[bcrypt]==1.7.4
